peddantic mode

+++ branch_predictor/btb_line.hpp
@@ -5,7 +5,7 @@ class btb_line_t{
 public:
     uint64_t tag;
     uint64_t lru;
-    uint32_t targetAddress;
+    uint64_t targetAddress;
     uint32_t validade;
     uint32_t typeBranch;
     uint32_t bht;
